requirements.txt
selenium
gspread
oauth2client
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
